Extract SourceFileConverter, reorganize test-utils into src
